Refactor AppRoutes
